requirements.txt
Django==1.7.4
djangorestframework==3.0.4
django-model-utils>=2.2,<2.3
gunicorn==19.2.0
dj-static==0.0.6
dj-database-url
psycopg2==2.5.4
google-api-python-client==1.3.1
pygeocoder 1.2.5
